Struct vizia_baseview::Application

pub struct Application<F>where
    F: Fn(&mut Context) + Send + 'static,{ /* private fields */ }

Implementations

impl<F> Application<F>where
F: Fn(&mut Context),
F: 'static + Send,

pub fn new(app: F) -> Self

pub fn ignore_default_theme(self) -> Self

pub fn with_scale_policy(self, scale_policy: WindowScalePolicy) -> Self

Change the window’s scale policy. Not part of new() to keep the same signature as the winit backend.

pub fn title(self, title: &str) -> Self

pub fn inner_size(self, size: impl Into<WindowSize>) -> Self

pub fn run(self)

Open a new window that blocks the current thread until the window is destroyed.

Do not use this in the context of audio plugins, unless it is compiled as a standalone application.

app - The Vizia application builder.

pub fn open_parented<P: HasRawWindowHandle>(self, parent: &P) -> WindowHandle

Open a new child window.

This function does not block the current thread. This is only to be used in the context of audio plugins.

parent - The parent window.

app - The Vizia application builder.

pub fn open_as_if_parented(self) -> WindowHandle

Open a new window as if it had a parent window.

This function does not block the current thread. This is only to be used in the context of audio plugins.

app - The Vizia application builder.

pub fn on_idle<I: 'static + Fn(&mut Context) + Send>(self, callback: I) -> Self

Takes a closure which will be called at the end of every loop of the application.

The callback provides a place to run ‘idle’ processing and happens at the end of each loop but before drawing. If the callback pushes events into the queue in context then the event loop will re-run. Care must be taken not to push events into the queue every time the callback runs unless this is intended.

Example

Application::new(|cx|{
    // Build application here
})
.on_idle(|cx|{
    // Code here runs at the end of every event loop after OS and vizia events have been handled
})
.run();

pub fn text_shaping_run_cache(self, size: usize) -> Self

Resize the cache used for rendering text lines

pub fn text_shaped_words_cache(self, size: usize) -> Self

Resize the cache used for rendering words